Add command update.
With the command `onepunch update` the user will update files in the `src` directory according to the latest release of onepunch.
Please note that any custom change inside directory `src` will be overwritten. In order to prevent accidental deletions, the command ckecks the presence of file `onepunch.json`, if the file is not present the command aborts.